+++ src/main/java/com/epam/emotionalhelp/controller/RecommendationController.java
@@ -1,11 +1,11 @@
 package com.epam.emotionalhelp.controller;
 
-import com.epam.emotionalhelp.model.RecommendationEntity;
+import com.epam.emotionalhelp.controller.dto.AppointmentResponseDto;
 import com.epam.emotionalhelp.service.RecommendationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,13 +19,8 @@ import java.util.List;
 public class RecommendationController {
     private final RecommendationService recommendationService;
 
-    @GetMapping("/attempt/{attemptId}")
-    public List<RecommendationEntity> findAllByAttemptId(@Min(1) @PathVariable Long attemptId) {
+    @PostMapping("/attempt/{attemptId}")
+    public List<AppointmentResponseDto> findAllByAttemptId(@Min(1) @PathVariable Long attemptId) {
         return recommendationService.findAllByAttemptId(attemptId);
-    }
-
-    @GetMapping("/{emotionId}")
-    public List<RecommendationEntity> findAllByEmotionId(@Min(1) @PathVariable Long emotionId) {
-        return recommendationService.findAllByEmotionId(emotionId);
     }
 }
