+++ src/main/java/com/epam/emotionalhelp/service/impl/RecommendationServiceImpl.java
@@ -1,9 +1,14 @@
 package com.epam.emotionalhelp.service.impl;
 
+import com.epam.emotionalhelp.controller.dto.AppointmentResponseDto;
 import com.epam.emotionalhelp.controller.dto.EmotionDto;
+import com.epam.emotionalhelp.model.AppointmentEntity;
+import com.epam.emotionalhelp.model.QuizResult;
 import com.epam.emotionalhelp.model.RecommendationEntity;
+import com.epam.emotionalhelp.repository.QuizResultRepository;
 import com.epam.emotionalhelp.repository.RecommendationRepository;
 import com.epam.emotionalhelp.service.RecommendationService;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,34 +21,39 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class RecommendationServiceImpl implements RecommendationService {
     private final RecommendationRepository recommendationRepository;
+    private final QuizResultRepository quizResultRepository;
 
     @Override
     @Transactional
-    public List<RecommendationEntity> findAllByAttemptId(Long attemptId) {
-        // have to get the list of EmotionDto by attemptId from DB ...
-        List<EmotionDto> dummyFake = List.of(EmotionDto.builder()
-                .emotionId(1L)
-                .value(10)
-                .build(), EmotionDto.builder()
-                .emotionId(6L)
-                .value(30)
-                .build());
-        return findForEmotionDtoList(dummyFake);
-    }
-
-    private List<RecommendationEntity> findForEmotionDtoList(List<EmotionDto> emotions) {
-        return emotions.stream()
+    public List<AppointmentResponseDto> findAllByAttemptId(Long attemptId) {
+        return quizResultRepository.findQuizResultsByAttemptId(attemptId).stream()
+                .map(Mapper::toEmotionDto)
                 .flatMap(em -> recommendationRepository.findAllByEmotionId(em.getEmotionId()).stream()
                         .filter(isInsideInterval(em)))
+                .map(RecommendationEntity::getAppointment)
+                .map(Mapper::toAppointmentResponseDto)
                 .collect(Collectors.toList());
+    }
+
+    private static class Mapper {
+        private static AppointmentResponseDto toAppointmentResponseDto(@NonNull AppointmentEntity ap) {
+            return AppointmentResponseDto.builder()
+                    .type(ap.getAppointmentType().getDescription())
+                    .icon(ap.getIcon())
+                    .description(ap.getDescription())
+                    .link(ap.getLink())
+                    .build();
+        }
+
+        private static EmotionDto toEmotionDto(@NonNull QuizResult qr) {
+            return EmotionDto.builder()
+                    .emotionId(qr.getEmotion().getId())
+                    .value(qr.getScore())
+                    .build();
+        }
     }
 
     private Predicate<RecommendationEntity> isInsideInterval(EmotionDto em) {
         return rec -> rec.getFloor() <= em.getValue() && em.getValue() <= rec.getCeil();
-    }
-
-    @Override
-    public List<RecommendationEntity> findAllByEmotionId(Long emotionId) {
-        return recommendationRepository.findAllByEmotionId(emotionId);
     }
 }
